Moved fido2 services and abstractions to vault folder in libs
