+++ libs/common/src/vault/abstractions/fido2/fido2-client.service.abstraction.ts
