+++ libs/common/src/vault/abstractions/fido2/fido2-user-interface.service.abstraction.ts
