+++ libs/common/src/vault/services/fido2/noop-fido2-user-interface.service.ts
@@ -1,7 +1,7 @@
 import {
   Fido2UserInterfaceService as Fido2UserInterfaceServiceAbstraction,
   Fido2UserInterfaceSession,
-} from "../abstractions/fido2-user-interface.service.abstraction";
+} from "../../abstractions/fido2/fido2-user-interface.service.abstraction";
 
 export class Fido2UserInterfaceService implements Fido2UserInterfaceServiceAbstraction {
   newSession(abortController?: AbortController): Promise<Fido2UserInterfaceSession> {
